# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Book Lending API's",
        "description": "How to login:\n *   Login using the `/login` endpoint with your email and password.\n *   Copy the token from the response, specifically from `data.token.value`.\n *   Click the Authorize button in the top right corner.\n *   Enter the token in the format: `Bearer <access_token>`.\n *   Click the Authorize button.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/",
            "description": "LIVE server"
        }
    ],
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "operationId": "5d4fa5d5607ed675ba93907e1fc94534",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login to system and return the token.",
                "description": "Do action login for access the API's",
                "operationId": "67d6d0fdd54fba132e1b6f55ca254183",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AuthUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get books with pagination",
                "description": "Get books with pagination",
                "operationId": "e7d6ea14835e1703a0d0c7ce68e8af6d",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "search for title or author",
                        "required": false,
                        "schema": {
                            "type": "sring"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "sort data by specific column (default by title)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "title",
                                "author",
                                "available_copies",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "description": "sort order (default by asc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Create a new book",
                "operationId": "fe4512256a41d7e8df654a7317ddf6ea",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateBook"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "CreateBook": {
                "required": [
                    "title",
                    "author",
                    "available_copies"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Some Book Title"
                    },
                    "author": {
                        "type": "string",
                        "example": "Some Author"
                    },
                    "available_copies": {
                        "type": "integer",
                        "example": 20
                    }
                },
                "type": "object"
            },
            "AuthUser": {
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "user1@mail.test"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
            "RegisterUser": {
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Your Name"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "yourmail@mail.test"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "format": "password",
                        "example": "123456"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "Bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Books",
            "description": "Books"
        }
    ]
}